\name{BinaryDosageFiles-package}
\alias{BinaryDosageFiles-package}
\alias{BinaryDosageFiles}
\docType{package}
\title{
\packageTitle{BinaryDosageFiles}
}
\description{
\packageDescription{BinaryDosageFiles}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BinaryDosageFiles}
\packageIndices{BinaryDosageFiles}
Functions to create and read dosage files in a binary format.
}
\author{
\packageAuthor{BinaryDosageFiles}

Maintainer: \packageMaintainer{BinaryDosageFiles}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
