% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExtractMoreDosages}
\alias{ExtractMoreDosages}
\title{Function to extract more SNPs from a binary dosage file}
\usage{
ExtractMoreDosages(inputs)
}
\arguments{
\item{inputs}{List of inputs returned from ExtractDosages}
}
\value{
List with a vector of dosages and a matrix of probabilities
and a list of the input values
}
\description{
Function to extract more SNPs from a binary dosage file
}
