% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExtractSNPDosages}
\alias{ExtractSNPDosages}
\title{Function to extract a SNP from a binary dosage file}
\usage{
ExtractSNPDosages(bdosageFilename, mapFilename, numSub, snpName, flanking)
}
\arguments{
\item{bdosageFilename}{Name of binary dosage file}

\item{mapFilename}{Name of map file associated with dosage file}

\item{numSub}{Number of subjects with data in dosage file}

\item{snpName}{Name of SNP to extract}

\item{flanking}{Number of flanking SNPs on either side to include}
}
\value{
A list
\item{Dosage}{Matrix of dosage values}
\item{P0}{Matrix of probabilities that g = 0}
\item{P1}{Matrix of probabilities that g = 1}
\item{P2}{Matrix of probabilities that g = 2}
\item{SNPInfo}{ Data frame with information about SNPs output }
}
\description{
Function to extract a SNP from a binary dosage file
}
