% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNPInfo.R
\name{SNP_Info}
\alias{SNP_Info}
\title{Function to return SNP information read by ExtractDosages'}
\usage{
SNP_Info(d, x)
}
\arguments{
\item{d}{List returned from ExtractDosages or ExtractMoreDosages}

\item{x}{Integer or vector of integers indicates SNPs to retrieve data on}
}
\value{
Data frame with chromosome, SNPName, location in base pairs, reference allele, and alternate allele
}
\description{
Function to return information on the SNPs read in by ExtractDosages
or ExtractMoreDosages
}
