% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{VCF_to_BinaryDosage}
\alias{VCF_to_BinaryDosage}
\title{Function to convert a VCF file to a binary dosage file for GxEScan}
\usage{
VCF_to_BinaryDosage(vcfFilename, outBaseFilename, initSub)
}
\arguments{
\item{vcfFilename}{Name of VCF file}

\item{outBaseFilename}{Base filename of output files. Three files are output.
A family file with extension .fam.
A map file with extension .bim.
A binary dosage file with extenstion .bdosage.}

\item{initSub}{Amount of memory to allocate for subjects. Should be the number of subjects.
Additional memory will be allocated if not enough is initially allocated.
Having to do this can slow down the speed of the routine because of memory
allocations/deallocations.}
}
\value{
0 success
1 failure
}
\description{
Function to convert a VCF file to a binary dosage file for GxEScan. 
The binary formatted file is smaller that the VCF file and reads in much quicker.
}
