% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExtractSNPDosages}
\alias{ExtractSNPDosages}
\title{Function to extract a SNP from a binary dosage file}
\usage{
ExtractSNPDosages(bdosageFilename, mapFilename, numSub, snpName, flanking)
}
\arguments{
\item{bdosageFilename}{Name of binary dosage file}

\item{mapFilename}{Name of map file associated with dosage file}

\item{numSub}{Number of subjects with data in dosage file}

\item{snpName}{Name of SNP to extract}

\item{flanking}{Number of flanking SNPs on either side to include}
}
\value{
List with a vector of dosages and a matrix of probabilities
and a list of the input values
}
\description{
Function to extract a SNP from a binary dosage file
}
